feat(Product Catalog): Added manage categories, editing, deleting and adding...

See merge request pap2024z-z22/pap2024z-z22!7

+++ backend/src/main/java/pap/backend/category/CategoryService.java
@@ -4,6 +4,8 @@ package pap.backend.category;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pap.backend.product.Product;
+import pap.backend.product.ProductRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,10 +13,12 @@ import java.util.Optional;
 @Service
 public class CategoryService {
     private final CategoryRepository categoryRepository;
+    private final ProductRepository productRepository;
 
     @Autowired
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, ProductRepository productRepository) {
         this.categoryRepository = categoryRepository;
+        this.productRepository = productRepository;
     }
 
     public List<Category> getCategories() {
@@ -41,6 +45,13 @@ public class CategoryService {
         if (!exists) {
             throw new IllegalStateException("category with id " + categoryId + " does not exist");
         }
+
+        List<Product> products = productRepository.findProductsByCategoryId(categoryId);
+
+        if (!products.isEmpty()) {
+            throw new IllegalStateException("category with id " + categoryId + " has products");
+        }
+
         categoryRepository.deleteById(categoryId);
     }
